#857 NPE with in-memory and idle messaging client
